moved components out of improve-accessibility-insights.tsx and updated related files
