+++ src/DetailsView/components/settings-panel/settings/telemetry/telemetry-settings.tsx
@@ -1,11 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import {
-    EnableTelemetrySettingDescription,
-    enableTelemetrySettingsPanelTitle,
-    LinkComponentDeps,
-} from 'content/settings/improve-accessibility-insights';
+import { enableTelemetrySettingsPanelTitle } from 'content/settings/improve-accessibility-insights';
 import * as React from 'react';
+import { EnableTelemetrySettingDescription } from '../../../../../common/components/enable-telemetry-setting-description';
+import { LinkComponentDeps } from '../../../../../common/components/telemetry-notice';
 import { NamedFC } from '../../../../../common/react/named-fc';
 import { GenericToggle } from '../../../generic-toggle';
 import { SettingsProps } from '../settings-props';
