+++ src/tests/unit/tests/DetailsView/components/settings-panel/settings/telemetry/telemetry-settings.test.tsx
@@ -4,10 +4,10 @@ import { shallow } from 'enzyme';
 import { Toggle } from 'office-ui-fabric-react/lib/Toggle';
 import * as React from 'react';
 import { Mock, Times } from 'typemoq';
+import { EnableTelemetrySettingDescription } from '../../../../../../../../common/components/enable-telemetry-setting-description';
 import { NewTabLink } from '../../../../../../../../common/components/new-tab-link';
 import { UserConfigMessageCreator } from '../../../../../../../../common/message-creators/user-config-message-creator';
 import { UserConfigurationStoreData } from '../../../../../../../../common/types/store-data/user-configuration-store';
-import { EnableTelemetrySettingDescription } from '../../../../../../../../content/settings/improve-accessibility-insights';
 import { GenericToggle } from '../../../../../../../../DetailsView/components/generic-toggle';
 import {
     TelemetrySettings,
