+++ src/tests/unit/tests/common/components/enable-telemetry-setting-description.test.tsx
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+import { shallow } from 'enzyme';
+import * as React from 'react';
+import { EnableTelemetrySettingDescription } from '../../../../../common/components/enable-telemetry-setting-description';
+import { NewTabLink } from '../../../../../common/components/new-tab-link';
+import { LinkComponentDeps } from '../../../../../common/components/telemetry-notice';
+
+describe('EnableTelemetrySettingDescription', () => {
+    it('renders with LinkComponentDeps', () => {
+        const deps: LinkComponentDeps = {
+            LinkComponent: NewTabLink,
+        };
+
+        const wrapper = shallow(<EnableTelemetrySettingDescription {...deps} />);
+
+        expect(wrapper.getElement()).toMatchSnapshot();
+    });
+});
